data_collection/requirements.txt
httpx>=0.27
python-dotenv>=1.0
pytest>=7.0
pytest-asyncio>=0.21
